feat(cast): support balance call overrides
Share state and block override parsing across eth_call-based commands. Apply overrides to ERC20 balance queries and reject them for native eth_getBalance.
