+++ crates/cast/src/cmd/call.rs
@@ -1,4 +1,7 @@
-use super::run::{fetch_contracts_bytecode_from_trace, fetch_contracts_bytecode_via_rpc};
+use super::{
+    call_overrides::CallOverrideOpts,
+    run::{fetch_contracts_bytecode_from_trace, fetch_contracts_bytecode_via_rpc},
+};
 use crate::{
     Cast,
     debug::handle_traces,
@@ -8,14 +11,11 @@ use crate::{
 };
 use alloy_ens::NameOrAddress;
 use alloy_network::{Network, NetworkTransactionBuilder, TransactionBuilder};
-use alloy_primitives::{
-    Address, B256, Bytes, TxKind, U256, hex,
-    map::{AddressHashMap, HashMap},
-};
+use alloy_primitives::{Bytes, TxKind, U256, hex, map::AddressHashMap};
 use alloy_provider::{Provider, ext::DebugApi};
 use alloy_rpc_types::{
     BlockId, BlockNumberOrTag, BlockOverrides,
-    state::{StateOverride, StateOverridesBuilder},
+    state::StateOverride,
     trace::geth::{
         CallConfig, GethDebugBuiltInTracerType, GethDebugTracerType, GethDebugTracingCallOptions,
         GethDebugTracingOptions, GethTrace,
@@ -53,13 +53,7 @@ use foundry_evm::{
     traces::{InternalTraceMode, SparsedTraceArena, TraceRequirements},
 };
 use foundry_wallets::WalletOpts;
-use regex::Regex;
-use std::{str::FromStr, sync::LazyLock};
-
-// matches override pattern <address>:<slot>:<value>
-// e.g. 0x123:0x1:0x1234
-static OVERRIDE_PATTERN: LazyLock<Regex> =
-    LazyLock::new(|| Regex::new(r"^([^:]+):([^:]+):([^:]+)$").unwrap());
+use std::str::FromStr;
 
 /// CLI arguments for `cast call`.
 ///
@@ -181,38 +175,8 @@ pub struct CallArgs {
     #[arg(long, visible_alias = "la")]
     pub with_local_artifacts: bool,
 
-    /// Override the accounts balance.
-    /// Format: "address:balance,address:balance"
-    #[arg(long = "override-balance", value_name = "ADDRESS:BALANCE", value_delimiter = ',')]
-    pub balance_overrides: Option<Vec<String>>,
-
-    /// Override the accounts nonce.
-    /// Format: "address:nonce,address:nonce"
-    #[arg(long = "override-nonce", value_name = "ADDRESS:NONCE", value_delimiter = ',')]
-    pub nonce_overrides: Option<Vec<String>>,
-
-    /// Override the accounts code.
-    /// Format: "address:code,address:code"
-    #[arg(long = "override-code", value_name = "ADDRESS:CODE", value_delimiter = ',')]
-    pub code_overrides: Option<Vec<String>>,
-
-    /// Override the accounts state and replace the current state entirely with the new one.
-    /// Format: "address:slot:value,address:slot:value"
-    #[arg(long = "override-state", value_name = "ADDRESS:SLOT:VALUE", value_delimiter = ',')]
-    pub state_overrides: Option<Vec<String>>,
-
-    /// Override the accounts state specific slots and preserve the rest of the state.
-    /// Format: "address:slot:value,address:slot:value"
-    #[arg(long = "override-state-diff", value_name = "ADDRESS:SLOT:VALUE", value_delimiter = ',')]
-    pub state_diff_overrides: Option<Vec<String>>,
-
-    /// Override the block timestamp.
-    #[arg(long = "block.time", value_name = "TIME")]
-    pub block_time: Option<u64>,
-
-    /// Override the block number.
-    #[arg(long = "block.number", value_name = "NUMBER")]
-    pub block_number: Option<u64>,
+    #[command(flatten)]
+    pub overrides: CallOverrideOpts,
 }
 
 #[derive(Debug, Parser)]
@@ -661,82 +625,14 @@ impl CallArgs {
         Ok(())
     }
 
-    /// Parse state overrides from command line arguments.
-    pub fn get_state_overrides(&self) -> eyre::Result<Option<StateOverride>> {
-        // Early return if no override set - <https://github.com/foundry-rs/foundry/issues/10705>
-        if [
-            self.balance_overrides.as_ref(),
-            self.nonce_overrides.as_ref(),
-            self.code_overrides.as_ref(),
-            self.state_overrides.as_ref(),
-            self.state_diff_overrides.as_ref(),
-        ]
-        .iter()
-        .all(Option::is_none)
-        {
-            return Ok(None);
-        }
-
-        let mut state_overrides_builder = StateOverridesBuilder::default();
-
-        // Parse balance overrides
-        for override_str in self.balance_overrides.iter().flatten() {
-            let (addr, balance) = address_value_override(override_str)?;
-            state_overrides_builder =
-                state_overrides_builder.with_balance(addr.parse()?, balance.parse()?);
-        }
-
-        // Parse nonce overrides
-        for override_str in self.nonce_overrides.iter().flatten() {
-            let (addr, nonce) = address_value_override(override_str)?;
-            state_overrides_builder =
-                state_overrides_builder.with_nonce(addr.parse()?, nonce.parse()?);
-        }
-
-        // Parse code overrides
-        for override_str in self.code_overrides.iter().flatten() {
-            let (addr, code_str) = address_value_override(override_str)?;
-            state_overrides_builder =
-                state_overrides_builder.with_code(addr.parse()?, Bytes::from_str(code_str)?);
-        }
-
-        type StateOverrides = HashMap<Address, HashMap<B256, B256>>;
-        let parse_state_overrides =
-            |overrides: &Option<Vec<String>>| -> Result<StateOverrides, eyre::Report> {
-                let mut state_overrides: StateOverrides = StateOverrides::default();
-
-                overrides.iter().flatten().try_for_each(|s| -> Result<(), eyre::Report> {
-                    let (addr, slot, value) = address_slot_value_override(s)?;
-                    state_overrides.entry(addr).or_default().insert(slot.into(), value.into());
-                    Ok(())
-                })?;
-
-                Ok(state_overrides)
-            };
-
-        // Parse and apply state overrides
-        for (addr, entries) in parse_state_overrides(&self.state_overrides)? {
-            state_overrides_builder = state_overrides_builder.with_state(addr, entries);
-        }
-
-        // Parse and apply state diff overrides
-        for (addr, entries) in parse_state_overrides(&self.state_diff_overrides)? {
-            state_overrides_builder = state_overrides_builder.with_state_diff(addr, entries)
-        }
-
-        Ok(Some(state_overrides_builder.build()))
+    /// Parses state overrides from command line arguments.
+    pub fn get_state_overrides(&self) -> Result<Option<StateOverride>> {
+        self.overrides.get_state_overrides()
     }
 
-    /// Parse block overrides from command line arguments.
-    pub fn get_block_overrides(&self) -> eyre::Result<Option<BlockOverrides>> {
-        let mut overrides = BlockOverrides::default();
-        if let Some(number) = self.block_number {
-            overrides = overrides.with_number(U256::from(number));
-        }
-        if let Some(time) = self.block_time {
-            overrides = overrides.with_time(time);
-        }
-        if overrides.is_empty() { Ok(None) } else { Ok(Some(overrides)) }
+    /// Parses block overrides from command line arguments.
+    pub fn get_block_overrides(&self) -> Result<Option<BlockOverrides>> {
+        self.overrides.get_block_overrides()
     }
 }
 
@@ -756,138 +652,10 @@ impl figment::Provider for CallArgs {
     }
 }
 
-/// Parse an override string in the format address:value.
-fn address_value_override(address_override: &str) -> Result<(&str, &str)> {
-    address_override.split_once(':').ok_or_else(|| {
-        eyre::eyre!("Invalid override {address_override}. Expected <address>:<value>")
-    })
-}
-
-/// Parse an override string in the format address:slot:value.
-fn address_slot_value_override(address_override: &str) -> Result<(Address, U256, U256)> {
-    let captures = OVERRIDE_PATTERN.captures(address_override).ok_or_else(|| {
-        eyre::eyre!("Invalid override {address_override}. Expected <address>:<slot>:<value>")
-    })?;
-
-    Ok((
-        captures[1].parse()?, // Address
-        captures[2].parse()?, // Slot (U256)
-        captures[3].parse()?, // Value (U256)
-    ))
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use alloy_primitives::{U64, address, b256, fixed_bytes};
-
-    #[test]
-    fn test_get_state_overrides() {
-        let call_args = CallArgs::parse_from([
-            "foundry-cli",
-            "--override-balance",
-            "0x0000000000000000000000000000000000000001:2",
-            "--override-nonce",
-            "0x0000000000000000000000000000000000000001:3",
-            "--override-code",
-            "0x0000000000000000000000000000000000000001:0x04",
-            "--override-state",
-            "0x0000000000000000000000000000000000000001:5:6",
-            "--override-state-diff",
-            "0x0000000000000000000000000000000000000001:7:8",
-        ]);
-        let overrides = call_args.get_state_overrides().unwrap().unwrap();
-        let address = address!("0x0000000000000000000000000000000000000001");
-        if let Some(account_override) = overrides.get(&address) {
-            if let Some(balance) = account_override.balance {
-                assert_eq!(balance, U256::from(2));
-            }
-            if let Some(nonce) = account_override.nonce {
-                assert_eq!(nonce, 3);
-            }
-            if let Some(code) = &account_override.code {
-                assert_eq!(*code, Bytes::from([0x04]));
-            }
-            if let Some(state) = &account_override.state
-                && let Some(value) = state.get(&b256!(
-                    "0x0000000000000000000000000000000000000000000000000000000000000005"
-                ))
-            {
-                assert_eq!(
-                    *value,
-                    b256!("0x0000000000000000000000000000000000000000000000000000000000000006")
-                );
-            }
-            if let Some(state_diff) = &account_override.state_diff
-                && let Some(value) = state_diff.get(&b256!(
-                    "0x0000000000000000000000000000000000000000000000000000000000000007"
-                ))
-            {
-                assert_eq!(
-                    *value,
-                    b256!("0x0000000000000000000000000000000000000000000000000000000000000008")
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn test_get_state_overrides_empty() {
-        let call_args = CallArgs::parse_from([""]);
-        let overrides = call_args.get_state_overrides().unwrap();
-        assert_eq!(overrides, None);
-    }
-
-    #[test]
-    fn test_get_block_overrides() {
-        let mut call_args = CallArgs::parse_from([""]);
-        call_args.block_number = Some(1);
-        call_args.block_time = Some(2);
-        let overrides = call_args.get_block_overrides().unwrap().unwrap();
-        assert_eq!(overrides.number, Some(U256::from(1)));
-        assert_eq!(overrides.time, Some(2));
-    }
-
-    #[test]
-    fn test_get_block_overrides_empty() {
-        let call_args = CallArgs::parse_from([""]);
-        let overrides = call_args.get_block_overrides().unwrap();
-        assert_eq!(overrides, None);
-    }
-
-    #[test]
-    fn test_address_value_override_success() {
-        let text = "0x0000000000000000000000000000000000000001:2";
-        let (address, value) = address_value_override(text).unwrap();
-        assert_eq!(address, "0x0000000000000000000000000000000000000001");
-        assert_eq!(value, "2");
-    }
-
-    #[test]
-    fn test_address_value_override_error() {
-        let text = "invalid_value";
-        let error = address_value_override(text).unwrap_err();
-        assert_eq!(error.to_string(), "Invalid override invalid_value. Expected <address>:<value>");
-    }
-
-    #[test]
-    fn test_address_slot_value_override_success() {
-        let text = "0x0000000000000000000000000000000000000001:2:3";
-        let (address, slot, value) = address_slot_value_override(text).unwrap();
-        assert_eq!(*address, fixed_bytes!("0x0000000000000000000000000000000000000001"));
-        assert_eq!(slot, U256::from(2));
-        assert_eq!(value, U256::from(3));
-    }
-
-    #[test]
-    fn test_address_slot_value_override_error() {
-        let text = "invalid_value";
-        let error = address_slot_value_override(text).unwrap_err();
-        assert_eq!(
-            error.to_string(),
-            "Invalid override invalid_value. Expected <address>:<slot>:<value>"
-        );
-    }
+    use alloy_primitives::U64;
 
     #[test]
     fn can_parse_call_data() {
@@ -914,10 +682,10 @@ mod tests {
             "0x123:0x1:0x1234",
         ]);
 
-        assert_eq!(args.balance_overrides, Some(vec!["0x123:0x1234".to_string()]));
-        assert_eq!(args.nonce_overrides, Some(vec!["0x123:1".to_string()]));
-        assert_eq!(args.code_overrides, Some(vec!["0x123:0x1234".to_string()]));
-        assert_eq!(args.state_overrides, Some(vec!["0x123:0x1:0x1234".to_string()]));
+        assert_eq!(args.overrides.balance_overrides, Some(vec!["0x123:0x1234".to_string()]));
+        assert_eq!(args.overrides.nonce_overrides, Some(vec!["0x123:1".to_string()]));
+        assert_eq!(args.overrides.code_overrides, Some(vec!["0x123:0x1234".to_string()]));
+        assert_eq!(args.overrides.state_overrides, Some(vec!["0x123:0x1:0x1234".to_string()]));
     }
 
     #[test]
@@ -943,16 +711,19 @@ mod tests {
         ]);
 
         assert_eq!(
-            args.balance_overrides,
-            Some(vec!["0x123:0x1234".to_string(), "0x456:0x5678".to_string()])
-        );
-        assert_eq!(args.nonce_overrides, Some(vec!["0x123:1".to_string(), "0x456:2".to_string()]));
-        assert_eq!(
-            args.code_overrides,
+            args.overrides.balance_overrides,
             Some(vec!["0x123:0x1234".to_string(), "0x456:0x5678".to_string()])
         );
         assert_eq!(
-            args.state_overrides,
+            args.overrides.nonce_overrides,
+            Some(vec!["0x123:1".to_string(), "0x456:2".to_string()])
+        );
+        assert_eq!(
+            args.overrides.code_overrides,
+            Some(vec!["0x123:0x1234".to_string(), "0x456:0x5678".to_string()])
+        );
+        assert_eq!(
+            args.overrides.state_overrides,
             Some(vec!["0x123:0x1:0x1234".to_string(), "0x456:0x2:0x5678".to_string()])
         );
     }
